+++ taxtastic/subcommands/new_database.py
@@ -16,8 +16,8 @@
 
 Download the current version of the NCBI taxonomy and load it into
 ``database_file`` as an SQLite3 database.  If ``database_file``
-already exists, it will fail and leave it untouched unless you specify
-``-x`` or ``--clobber``.  The NCBI taxonomy will be downloaded into
+already exists it be will overwritten unless you specify ``--no-clobber``.
+The NCBI taxonomy will be downloaded into
 the same directory as ``database_file`` will be created in unless you
 specify ``-p`` or ``--download-dir``.
 """